fix: handle the case where there is only 1 philosopher
So far, this case triggered a dead-lock, because the philosopher
tries to lock its first fork, and then its second one, which is
the same as its first, and so is already locked by itself.
Now, the philosopher routine starts by checking that the 2 forks
are different, and if not, that means that there is only 1 philosopher
in the simulation, and so it announces taking a fork (actually,
it does not take it), then it waits for its time to die, and
finally it announces its death, ending the simulation properly.