perf: take the runtime latency into account
When a philosopher does an action, it now suspends the thread
for as long microseconds as needed to reach the next theorical
time when the philosopher should end its action if there was
no any runtime latency. For example, assuming that `time_to_eat`
is 200, and `time_to_sleep` is 100, a philosopher starting eating
at 0 and ending eating at 201 (1 millisecond of runtime latency)
will sleep until 300, not 301.